Core/CPU: Fix LEA returning wrong offsets

+++ Source/Core/CPU/CPU.cpp
@@ -412,7 +412,7 @@ void Tick()
     u16& dst16 = ParameterTo<u16&>(dst, ins.GetPrefix());
     u16& src16 = ParameterTo<u16&>(src, ins.GetPrefix());
 
-    dst16 = static_cast<u16>(&src16 - Memory::GetPtr<u16>(DS, 0));
+    dst16 = static_cast<u16>((intptr_t)&src16 - (intptr_t)Memory::GetPtr<u16>(DS, 0));
 
     break;
   }
